Store the key for a name to sid cache entry in upper case rather than whatever case the request was made in.  This gets rid of duplicate cache entries.
Also when doing a sid to name, prime the cache with the name to sid
mapping result.  We can't do the reverse as we don't know the correct
case of the name to store in the cache.
(This used to be commit f268b0d5fb811b364578b11a66ca69973717eea8)