Eeek - wrong commit message for the previous commit.  )-:
Added a flags uint32 to the internal print notify message struct.
Defined a flag to denote that a jobid number is a unix jobid, not a
smb jobid.

The print notify message handler now checks this flag and converts the
unix jobid to a smb jobid before sending the notify.  The port monitor
gets very confused if you send it a notification for a nonexistent job.