the last WINS update broke self registration when we are a WINS server. The real problem is all the special cases we had for when we are a wins server as opposed to when we are using a 'real' wins server.
This patch removes the special cases. We now accept non-broadcast
packets from ourselves and we use ourselves as a wins server when we
are one. This gets rid of the special cases and simplifies things
quite a bit.

It all seems to work, but there are bound to be problems found later.
(This used to be commit 3e843d30158d05cdfba716bac7e5c0a75ae7a79c)