Fix up a botched prevoius commit.
The idea here is to allow invalid LM passwords in otherwise valid accounts.

This happens when we create an account without a password, for example.
Previously we would stop at the LM password, and not read things like the
account flags correctly.  Now we process the record, and just set the password
to NULL.

(Note, 'no password for access' is decided only on the basis of the Account
Control bits, not on the 'NULL' value of the password feild.).

Andrew Bartlett