added LDAP_SET_REBIND_PROC_ARGS in acconfig.h
andrew, you seem to have added this test but don't use it. Do you
intend to use it later? If not then perhaps it can be removed.

also, when a test goes in configure.in you must also add it to
acconfig.h, or you end up breaking configure.
(This used to be commit 496cd0876cc13e2dd25c6ddbfe04c5787dddb4dd)