These pointers should be for the service we just cloned, not the new service as it might actually not be a clone, but a 'found' service.
Previously this meant that a second session setup on the same TCP/IP connection
for a user would write '' as the home directory.  Instead, I want it to update
the share to be there current home directory (so that this can be updated
without the need to reboot the client for example).

Andrew Bartlett