Fix a bug in the netbios name cache code pointed out by Jerry.  It was possible for a lucky non-root user to initialise the namecache which, due to the permissions of the lock file directory, would probably fail.  Thus the namecache would not work properly.
We now initialise the namecache as root in smbd after the fork, and in
the initialisation section of winbindd.

Client daemons need to call namecache_enable() somewhere in their
initialisation to use the cache.  Otherwise, the namecache is not
used.  This avoids the induction of another member into the "make sure
you call this magic initialisation function or wierd things will
happen club".