Patch from mimir to back out idra's attempted DOS mitigation patch. (It broke port 139 name exchange)
I've been thinking about this, and doing is properly is actually rather
difficult - but I'll try and get somthing in there.  (My worry is what
smb_read_error should be set to, and how that interacts with the rest of
samba).

Andrew Bartlett
(This used to be commit 3e682867bbb13dae265cb9a8acea8b7cc87d82a6)