Fix bug when using smbclient to connect to a Win98 file share. The first character of the filename was being truncated.
Jeremy, This looks like a bug in the understanding of the
SMB_INFO_STANDARD (info level 1) for a FIND_FIRST2 call.  Please
check this.  It does not match what is stated in the SNIA spec
but works correctly against Win98se.

The spec says that attributes should be a USHORT and we parse it as
a UCHAR.