When compiled --with-ldapsam, make ldapsam the default passdb backend.
This is to allow painless upgrades from 2.2, and so people don't get a shock
when they follow old docs.

If ldap has been detected on the system, ldapsam is always available, just not
the default.

Andrew Bartlett
(This used to be commit 0a6a0c88d0972fcea4aead7115929f96c0d23cbc)