Back out one of the API changes, now I recall how it was meant to work.
(New accounts are created on thier internal talloc context, not a supplied one
and not the one beloning to the SAM)

Also actaully add the sam_skel to the tree.

I need to work out with Jelmer an appropriate way to ensure we still have an
example external build, and if this should move back out again.

Andrew Bartlett
(This used to be commit dc8b0d23e088d1e7f7fbc6ab0423eaa82f14e258)