Back out one of the API changes, now I recall how it was meant to work.
(New accounts are created on thier internal talloc context, not a supplied one
and not the one beloning to the SAM)

Also actaully add the sam_skel to the tree.

I need to work out with Jelmer an appropriate way to ensure we still have an
example external build, and if this should move back out again.

Andrew Bartlett