Another patch from Steve Langasek <vorlon@netexpress.net>, again from the Debian patchset.
The idea is to still allow DNS domains to be specified in 'hosts allow' and
'hosts deny' without making the admin set 'hostname lookups' in their smb.conf.
His concern is about upgrades.

This has been designed not to change the value of %M.

Andrew Bartlett
(This used to be commit f428ee98d83a55922c991b0eeddd47891b80d918)