--------------                       /              \                      /      REST      \                     /        IN        \                    /       PEACE        \                   /                      \                   |    global_myname     |                   |  global_myworkgroup  |                   |                      |                   |      30 October      |                   |                      |                   |         2002         |                   |                      |                   | "You will not be     |                   |    missed"           |                   |                      |                  *|     *  *  *          | *         _________)/\\_//(\/(/\)/\//\/\///|_)_______
Made explicit when we're asking for a dos string, or
a UNIX string. We're still broken w.r.t. codepage issues,
but now hopefully we have a chance to fix it and we're
no worse than before.

Look for the elements of this code in lib/util.c to be
added to 3.0.x and HEAD soon.

Jeremy.

Attempt to make progress on CR#663. Needs *extensive*
testing.