Fixes for cr1044.  This changes a number of smb.conf parameters from P_STRING types to P_LIST types so that larger numbers of users can be made in to printer admins.  Previously we were limited to 1024 characters of admin users.
 - Changed read/write list,  valid/invalid users and printer admin
   parameters to be P_LIST types.

 - Merged across the user_in_list() equivalent from HEAD.  Luckily
   most of the P_LIST code has already been merged across for the WINS
   list code.

 - Made lp_winbind_separator() a FN_GLOBAL_CONST_STRING to speed up
   access checking.

 - Added another cached access check to the connection struct.  We now
   cache whether a user is a printer admin at connection time to avoid
   making a potentially long access check at every printer rpc.

The line length problem is fixed but we don't scale much past 100
or so users as in some cases we need to check every name in the list
to see whether it is a domain group and whether the user is a member
of it.

A better long term solution would be to make a security descriptor for
all these settings and just add/remove SID entries to it.