Instead of walking the entire group database, grabbing all members of each group, testing for membership etc, use the already calculated NT_USER_TOKEN.
(which is initgroups() based)

So far we only fill out the 'domain' groups - we need to fill out the
'other sids' as well, and we possibly need to filter the list for 'domain
groups' only (the old code did that), but Win2k doesn't seem to mind
this for now.

I also need to find out what the magic '7' is about.  Fortunetly JF is in town,
so I'll grill him tomorrow :-).

Andrew Bartlett
(This used to be commit 7e846e7387247f8bc2b07a1a1006014c6978143b)