Add a new VFS module, that just fiddles the file permissions.  Still need to make it fiddle with NT ACLs (or report no ACL, so Samba fakes it up).
The idea here is that Samba reports an NT ACL to Windows clients, which use
that ACL in downloaded profiles.

If the user doesn't have write permissions on the directories being downloaded,
then it cant put the subfolders in (yes, NT will set perms so that it can't
access the dir as the user the created it) and the profile download fails.

The current solution it to give the user unix write perms to the folders, but
this is rather dangerous - sombody could trojen the profile.  This should
avoid that mess.

I'll test this out properly next week, but it works in vfstest (thanks idra!).

Andrew Bartlett
(This used to be commit 1ad782282a3eb4f89c489a77ad8404099d41d0f9)