Don't initialise the trusted domain list on startup in winbindd_setup_common() This is now done on demand as all access to the list of domains is done through the domain_list() function.
This was a throwback to when winbindd had a server_state structure and
the domain SID was required to be known before winbindd could start
processing requests.  Now since the domain SID is stored in the domain
list which is read on demand we don't need to do this initialisation
right now.

Final part of CR1340.