Patch from ab, to make findsmb omit -r most of the time.
(original mail follows)
--
From:	Alexander Bokovoy <a.bokovoy@sam-solutions.net>
Greetings!

I propose to make an optional argument '-r' for findsmb utility to pass
'-r' option to nmblookup optionally and default to omit it. The reason is
to have findsmb working more user-friendly in contemporary office
environments where Windows 95/98 boxes are rare and also to allow usage of
the utility under non-priviledged accounts when Windows95/98 boxes are in
minority or even absent in network neighbourhood.

An attached patch provides this change for HEAD together with
documentation changes.

--
/ Alexander Bokovoy