(missed in last commit)
Change the 'cookie' to be the ntlmssp_context, and use the 'auth_context' on
that to store the cookie.  Ensures that simple callbacks can 'just work'.

Also make it clear that we are doing a pull_string into a pstring, not just
any sized buffer.

Andrew Bartlett
(This used to be commit c7793f27188e658b7fc6336aa51d367eab36fc17)