hp CR1548, CR1591
Fix nasty string pointer bug in global parameter table.  Basically we
were doing a[1][0] when we wanted a[0][1].  This caused messages like
"getpwnam /usr/bin/passwd", and also the server description not to
show up in network neighbourhood.

This fixes code that only exists in appliance_head (not 2.2, 3.0 or
head), and so shouldn't need to be ported as far as I can see.