closing bug PR#26498
wide links parameter was not working for symlinks to files
(see PR#26498 for the discussion).

I tried to be as less invasive as possible in reduce_name function so that I should have not changed any previous behaviour but only addedd the missing check.

Jeremy can you check thi is ok?

Simo.